requirements.txt
torch
tensorboard
scikit-learn
pyyaml
matplotlib
torchinfo